Add `Details` component w/ dynamic content
In an unconventional move, I'm storing the text for the component
as an array of static JSX blocks, and inserting the one that matches
the route on render.

Would not scale well, but works great for my purposes. No need for a
database or a bunch of markdown in external files =)

Also updated flipMove temporarily.