Throttle and refine Square demo
The Fuscia Square demo was buggy and awkward. To improve things, I
wrote a quick throttle function that ignores multiple requests, that
way the square's movement is predictable and never moves more than 1
place in the timeframe specified.

For organization, I moved this function into a new `helpers` dir,
along with the monkeypatched Array methods `range` and `matrix`.