Fix bug with destroyed children
In the last patch, I improved filtering to check whether DOM nodes
had moved, before moving on to the 'animation' step. This had the
inadvertant effect of always trying to get the bounding rects for
all DOM nodes in the previous state.

An exception was being thrown when DOM nodes were destroyed, because
it was trying to calculate the bounding rects for elements that no
longer exist.

Simple fix, by returning earlier if the DOM node is destroyed, new,
or immovable (no child key).