Add support for `className` & `typeName`
FlipMove works by enclosing the children it's wrapped around in an
ordinary, no-attribute <div>. Someone messaged me asking if I could
add the ability to supply a custom `className`, so that he could add
a `row` class to the container, so that it would play nice with
Foundation.

While I was there, I also added a `typeName` field, which allows
you to select a different element than `div`. So, if your children
are <li>s, you might want to supply 'ul'.