Enter/Exit Animations implementation
Rewrote massive chunks of FlipMove to store children in state. This
way, when new props are received, we can "hang on" to items that are
on the way out, to animate their exit.

The big hurdle with this is that if the item remains in the DOM, the
siblings don't know to ignore it in their layout calculations:

-----------         -----------
|  item 1 |         |  item 1 |
-----------         -----------
-----------         -----------
|  item 2 |    ->   |/////////|  <-
-----------         -----------    \
-----------         -----------    /
|  item 3 |         |  item 3 |  -
-----------         -----------

In this example, new props are received, and item 2 no longer exists.
We want to simultaneously animate item 2's exit (fade out, whatever),
while also moving item 3 up to take item 2's place.

The rendering engine, though, doesn't realize that item 2 has left;
it's still there, it's just fading out. So it can't recalculate the
layout to move item 3 up.

My solution is to take item 2 out of the document flow by giving it
an `absolute` positioning, and matching its on-screen location using
bounding boxes.

Another challenge involved interruptions. Ensuring that if an item
is mid-exit and new props come in (that may or may not include that
same item, not exiting), nothing funky happens.

Finally, I added some presets, and the structure to parse presets
into plain styles.

Whew. Big commit.