Add configuration and test coverage
Add a boolean that will enable the adding of a placeholder element to hold open the container while elements are "leaving" (and therefore removed from the flow).

Add a single test case that checks that the container size is maintained immediately after element is removed.

Add story book entry demonstrating the "maintain container height" option's behaviour.

Closes joshwcomeau/react-flip-move#80