minor change to example custom animation in readme
I liked this custom animation a lot, but found that it caused weird 'hangs' at the beginning and end of animations when sorting multiple items per transition. By making these changes, the transitions are greatly smoothed out and much more pleasing to the eye. I also changed the format just a little so that its more common usage pattern is a little more obvious.